requirements.txt
gradio
torch
torchvision
Pillow
numpy